requirements.txt
pytest
python-dotenv
requests
langchain-community
langchain-ollama
langchain_ollama
googlesearch-python
beautifulsoup4

